refactor: tidy CodeEditor types, theme, and variants
- Reuse the shared InputVariant union (via Exclude<…, 'ghost'>) instead of
  redeclaring it, so the code field can't drift from the inputs beside it.
- Drop the ghost variant: a borderless code surface reads as plain text and
  loses the affordance that it's an editable field. Removed from the type, the
  theme fragment, the variants story, the playground knob, and the docs.
- Type language as CodeLanguage | (string & {}) to keep autocomplete while
  still allowing arbitrary keys.
- Extract the ~200 lines of theme + highlightStyle literals into a sibling
  theme.ts (taking the lazy-imported CodeMirror modules as params so the
  static graph stays editor-free).
