+++ src/components/CodeEditor/stories/Variants.vue
@@ -9,6 +9,5 @@ const code = ref(`{\n  "name": "frappe-ui",\n  "private": true\n}`)
   <div class="w-full max-w-xl space-y-4">
     <CodeEditor v-model="code" language="json" variant="subtle" />
     <CodeEditor v-model="code" language="json" variant="outline" />
-    <CodeEditor v-model="code" language="json" variant="ghost" />
   </div>
 </template>
